+++ NEWS.md
@@ -1,6 +1,8 @@
 # hammr 1.0.0.9006
 
 * added dsinfo: set and display information about a data set
+* added df_value_replace_if: replace a value in data frame columns that match
+  a predicate